# internal/client/generated/api/openapi.yaml
openapi: 3.0.3
info:
  description: |-
    Ory Talos is a high-performance API key management service. It handles the full API key lifecycle: issuing keys, verifying them at low latency, deriving short-lived tokens (JWT and Macaroon), and revoking access.

    `Admin*`-prefixed RPCs require admin authentication. The bare `RevokeApiKey` RPC is authenticated by proof of possession (the raw credential secret).
  title: Ory Talos API
  version: v2alpha1
servers:
- url: /
...
paths:
...
  /v2alpha1/admin/importedApiKeys:
    get:
      description: |-
        Lists all imported keys with filtering. Returns imported keys only (not
        issued keys). Supports pagination and AIP-160 filter expressions.

        ```http
        GET /v2alpha1/admin/importedApiKeys?page_size=50&filter=status%3DKEY_STATUS_ACTIVE
        ```
      operationId: adminListImportedApiKeys
      parameters:
      - description: "Number of items per page (default: 50, max: 1000)"
        explode: true
        in: query
        name: page_size
        required: false
        schema:
          format: int32
          type: integer
        style: form
      - description: Cursor token for pagination (OPTIONAL)
        explode: true
        in: query
        name: page_token
        required: false
        schema:
          type: string
        style: form
      - description: |-
          filter is an AIP-160 expression. Indexed fields (efficient at any scale):
            actor_id, status. Other fields are not indexed and may be rejected.
          Examples:
            actor_id="user_123"
            status=KEY_STATUS_ACTIVE
            actor_id="user_123" AND status=KEY_STATUS_ACTIVE
        explode: true
        in: query
        name: filter
        required: false
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListImportedApiKeysResponse"
          description: A successful response.
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Status"
          description: An unexpected error response.
      summary: List Imported API Keys
      tags:
      - ApiKeys
    post:
      description: |-
        Imports an external API key into the system. Allows importing keys from
        legacy systems or external providers. The raw key is hashed and stored
        securely (HMAC). Imported keys support token derivation (JWT/Macaroon)
        like issued keys.

        ```http
        POST /v2alpha1/admin/importedApiKeys
        {
          "raw_key": "sk_live_abc123xyz",
          "name": "Imported Stripe Key",
          "actor_id": "user_123"
        }
        ```
      operationId: adminImportApiKey
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportApiKeyRequest"
        description: |-
          Example:
            {
              "raw_key": "sk_live_abc123xyz789",
              "name": "Stripe Production Key",
              "actor_id": "payment-processor",
              "scopes": ["read", "write"],
              "ttl": "8760h",  // 1 year (also accepts: 31536000s)
              "metadata": {"source": "stripe", "environment": "production"}
            }
        required: true
        x-originalParamName: body
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportedApiKey"
          description: A successful response.
        "201":
          content:
            application/json:
              schema: {}
          description: API key imported successfully.
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Status"
          description: An unexpected error response.
      summary: Import API Key
      tags:
      - ApiKeys
...
  /v2alpha1/admin/issuedApiKeys:
    get:
      description: |-
        Lists issued API keys with optional filtering. Supports cursor-based
        pagination and AIP-160 filter expressions. Returns only issued
        (generated) API keys; use ListImportedApiKeys for imported keys.

        ```http
        GET /v2alpha1/admin/issuedApiKeys?page_size=50&filter=actor_id%3D%22user_123%22
        ```
      operationId: adminListIssuedApiKeys
      parameters:
      - description: "Number of items per page (default: 50, max: 1000)"
        explode: true
        in: query
        name: page_size
        required: false
        schema:
          format: int32
          type: integer
        style: form
      - description: Cursor token for pagination
        explode: true
        in: query
        name: page_token
        required: false
        schema:
          type: string
        style: form
      - description: |-
          filter is an AIP-160 expression. Indexed fields (efficient at any scale):
            actor_id, status. Other fields are not indexed and may be rejected.
          Examples:
            actor_id="user_123"
            status=KEY_STATUS_ACTIVE
            actor_id="user_123" AND status=KEY_STATUS_ACTIVE
        explode: true
        in: query
        name: filter
        required: false
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListIssuedApiKeysResponse"
          description: A successful response.
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Status"
          description: An unexpected error response.
      summary: List Issued API Keys
      tags:
      - ApiKeys
    post:
      description: |-
        Creates a new API key for a given actor. The secret is returned only once
        in the response and cannot be retrieved later. Keys can be scoped with
        specific permissions and have optional expiration.

        ```http
        POST /v2alpha1/admin/issuedApiKeys
        {
          "name": "production-service",
          "actor_id": "user_123",
          "scopes": ["read", "write"],
          "ttl": "8760h"
        }
        ```
      operationId: adminIssueApiKey
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IssueApiKeyRequest"
        required: true
        x-originalParamName: body
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssueApiKeyResponse"
          description: A successful response.
        "201":
          content:
            application/json:
              schema: {}
          description: API key issued successfully.
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Status"
          description: An unexpected error response.
      summary: Issue API Key
      tags:
      - ApiKeys
...
  /v2alpha1/apiKeys:selfRevoke:
    post:
      description: |-
        Proof-of-possession variant of revocation. The `Self*` prefix on the
        request/response messages disambiguates from the admin variants
        (`AdminRevokeIssuedApiKey` / `AdminRevokeImportedApiKey`).

        Allows an API key holder to revoke their own key. The caller must provide
        the full API key secret as proof of possession. Supports issued API keys
        and imported keys. JWT and macaroon tokens cannot be self-revoked (they
        are stateless).

        The PRIVILEGE_WITHDRAWN reason is not allowed for self-revocation
        (admin-only).

        ```http
        POST /v2alpha1/apiKeys:selfRevoke
        {
          "credential": "sk_live_abc123...",
          "reason": "REVOCATION_REASON_KEY_COMPROMISE"
        }
        ```
      operationId: revokeApiKey
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SelfRevokeApiKeyRequest"
        description: |-
          SelfRevokeApiKeyRequest allows an API key holder to revoke their own key
          by providing the full key secret as proof of possession.
        required: true
        x-originalParamName: body
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SelfRevokeApiKeyResponse"
          description: A successful response.
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Status"
          description: An unexpected error response.
      summary: Revoke API Key (self-service)
      tags:
      - ApiKeys
...
components:
  schemas:
...
    Any:
      additionalProperties: {}
      example:
        '@type': '@type'
      properties:
        '@type':
          type: string
      type: object
...
    ImportApiKeyRequest:
      description: |-
        Example:
          {
            "raw_key": "sk_live_abc123xyz789",
            "name": "Stripe Production Key",
            "actor_id": "payment-processor",
            "scopes": ["read", "write"],
            "ttl": "8760h",  // 1 year (also accepts: 31536000s)
            "metadata": {"source": "stripe", "environment": "production"}
          }
      example:
        metadata: "{}"
        rate_limit_policy:
          unit: unit
          quota: quota
          window: window
        visibility: KEY_VISIBILITY_UNSPECIFIED
        name: name
        actor_id: actor_id
        scopes:
        - scopes
        - scopes
        request_id: request_id
        ttl: ttl
        ip_restriction:
          allowed_cidrs:
          - allowed_cidrs
          - allowed_cidrs
        raw_key: raw_key
      properties:
        actor_id:
          description: |-
            actor_id is the identifier of the entity that owns this imported key.
            Required so every imported key is traceable to an actor for revocation and
            audit queries.
          type: string
        ip_restriction:
          $ref: "#/components/schemas/IPRestriction"
        metadata:
          description: |-
            metadata is a free-form JSON object for caller-defined attributes (e.g.,
            source, environment, tags). Values may be strings, numbers, booleans,
            arrays, objects, or null. Total serialized size is capped at 4KB.
            AIP-148 metadata field.
          type: object
        name:
          title: Human-readable name (REQUIRED)
          type: string
        rate_limit_policy:
          $ref: "#/components/schemas/RateLimitPolicy"
        raw_key:
          title: The actual key string to import (REQUIRED)
          type: string
        request_id:
          title: Client-controlled idempotency key (AIP-155)
          type: string
        scopes:
          items:
            type: string
          title: Scopes for the imported key (OPTIONAL)
          type: array
        ttl:
          description: |-
            ttl sets the expiry as a duration from now. Encoded as a
            google.protobuf.Duration (string ending in "s", e.g. "3600s").
            Accepted bounds: 1s to 315360000s (~10 years). If unset or zero, the
            project default TTL applies.
            For convenience, the server also accepts Go-style duration strings
            ("24h", "30m", "1h30m") and an extended unit set ("1d", "1w", "1mo",
            "1y"; approximations: 1mo = 30d, 1y = 365d). Clients should prefer
            the standard Duration encoding for portability.
          type: string
        visibility:
          $ref: "#/components/schemas/KeyVisibility"
      title: ImportApiKeyRequest imports an external HMAC-based API key
      type: object
    ImportedApiKey:
      description: |-
        ImportedApiKey represents an API key imported from an external system.
        The raw key is hashed (SHA-512/256) and stored. The original key is never retained.
      example:
        metadata: "{}"
        create_time: 2000-01-23T04:56:07.000+00:00
        rate_limit_policy:
          unit: unit
          quota: quota
          window: window
        visibility: KEY_VISIBILITY_UNSPECIFIED
        key_id: key_id
        expire_time: 2000-01-23T04:56:07.000+00:00
        revocation_description: revocation_description
        update_time: 2000-01-23T04:56:07.000+00:00
        last_used_time: 2000-01-23T04:56:07.000+00:00
        name: name
        revocation_reason: REVOCATION_REASON_UNSPECIFIED
        actor_id: actor_id
        scopes:
        - scopes
        - scopes
        ip_restriction:
          allowed_cidrs:
          - allowed_cidrs
          - allowed_cidrs
        status: KEY_STATUS_UNSPECIFIED
      properties:
        actor_id:
          type: string
        create_time:
          format: date-time
          type: string
        expire_time:
          format: date-time
          type: string
        ip_restriction:
          $ref: "#/components/schemas/IPRestriction"
        key_id:
          title: SHA-512/256 hash of credential
          type: string
        last_used_time:
          format: date-time
          type: string
        metadata:
          description: |-
            metadata is a free-form JSON object for caller-defined attributes
            (e.g., source, environment, tags). Values may be strings, numbers,
            booleans, arrays, objects, or null. Total serialized size is capped
            at 4KB. AIP-148 metadata field.
          type: object
        name:
          type: string
        rate_limit_policy:
          $ref: "#/components/schemas/RateLimitPolicy"
        revocation_description:
          description: |-
            revocation_description provides free-form context for a revocation.
            Only set when revocation_reason is PRIVILEGE_WITHDRAWN.
            JSON API change: field was formerly revocation_reason_text. Field number 13
            is unchanged so the change is wire-compatible for binary proto encoding.
          type: string
        revocation_reason:
          $ref: "#/components/schemas/RevocationReason"
        scopes:
          items:
            type: string
          type: array
        status:
          $ref: "#/components/schemas/KeyStatus"
        update_time:
          format: date-time
          type: string
        visibility:
          $ref: "#/components/schemas/KeyVisibility"
      type: object
    IssueApiKeyRequest:
      example:
        metadata: "{}"
        rate_limit_policy:
          unit: unit
          quota: quota
          window: window
        visibility: KEY_VISIBILITY_UNSPECIFIED
        name: name
        actor_id: actor_id
        scopes:
        - scopes
        - scopes
        request_id: request_id
        ttl: ttl
        ip_restriction:
          allowed_cidrs:
          - allowed_cidrs
          - allowed_cidrs
      properties:
        actor_id:
          type: string
        ip_restriction:
          $ref: "#/components/schemas/IPRestriction"
        metadata:
          description: |-
            metadata is a free-form JSON object for caller-defined attributes
            (e.g., source, environment, tags). Values may be strings, numbers,
            booleans, arrays, objects, or null. Total serialized size is capped
            at 4KB. AIP-148 metadata field.
          type: object
        name:
          type: string
        rate_limit_policy:
          $ref: "#/components/schemas/RateLimitPolicy"
        request_id:
          title: Client-controlled idempotency key (AIP-155)
          type: string
        scopes:
          items:
            type: string
          type: array
        ttl:
          description: |-
            ttl sets the expiry as a duration from now. Encoded as a
            google.protobuf.Duration (string ending in "s", e.g. "3600s").
            Accepted bounds: 1s to 315360000s (~10 years). If unset or zero, the
            project default TTL applies.
            For convenience, the server also accepts Go-style duration strings
            ("24h", "30m", "1h30m") and an extended unit set ("1d", "1w", "1mo",
            "1y"; approximations: 1mo = 30d, 1y = 365d). Clients should prefer
            the standard Duration encoding for portability.
          type: string
        visibility:
          $ref: "#/components/schemas/KeyVisibility"
      type: object
    IssueApiKeyResponse:
      example:
        secret: secret
        issued_api_key:
          metadata: "{}"
          create_time: 2000-01-23T04:56:07.000+00:00
          rate_limit_policy:
            unit: unit
            quota: quota
            window: window
          visibility: KEY_VISIBILITY_UNSPECIFIED
          key_id: key_id
          expire_time: 2000-01-23T04:56:07.000+00:00
          revocation_description: revocation_description
          update_time: 2000-01-23T04:56:07.000+00:00
          last_used_time: 2000-01-23T04:56:07.000+00:00
          name: name
          revocation_reason: REVOCATION_REASON_UNSPECIFIED
          actor_id: actor_id
          scopes:
          - scopes
          - scopes
          ip_restriction:
            allowed_cidrs:
            - allowed_cidrs
            - allowed_cidrs
          status: KEY_STATUS_UNSPECIFIED
      properties:
        issued_api_key:
          $ref: "#/components/schemas/IssuedApiKey"
        secret:
          title: Only returned on creation
          type: string
      type: object
...
    ListImportedApiKeysResponse:
      example:
        next_page_token: next_page_token
        imported_api_keys:
        - metadata: "{}"
          create_time: 2000-01-23T04:56:07.000+00:00
          rate_limit_policy:
            unit: unit
            quota: quota
            window: window
          visibility: KEY_VISIBILITY_UNSPECIFIED
          key_id: key_id
          expire_time: 2000-01-23T04:56:07.000+00:00
          revocation_description: revocation_description
          update_time: 2000-01-23T04:56:07.000+00:00
          last_used_time: 2000-01-23T04:56:07.000+00:00
          name: name
          revocation_reason: REVOCATION_REASON_UNSPECIFIED
          actor_id: actor_id
          scopes:
          - scopes
          - scopes
          ip_restriction:
            allowed_cidrs:
            - allowed_cidrs
            - allowed_cidrs
          status: KEY_STATUS_UNSPECIFIED
        - metadata: "{}"
          create_time: 2000-01-23T04:56:07.000+00:00
          rate_limit_policy:
            unit: unit
            quota: quota
            window: window
          visibility: KEY_VISIBILITY_UNSPECIFIED
          key_id: key_id
          expire_time: 2000-01-23T04:56:07.000+00:00
          revocation_description: revocation_description
          update_time: 2000-01-23T04:56:07.000+00:00
          last_used_time: 2000-01-23T04:56:07.000+00:00
          name: name
          revocation_reason: REVOCATION_REASON_UNSPECIFIED
          actor_id: actor_id
          scopes:
          - scopes
          - scopes
          ip_restriction:
            allowed_cidrs:
            - allowed_cidrs
            - allowed_cidrs
          status: KEY_STATUS_UNSPECIFIED
      properties:
        imported_api_keys:
          items:
            $ref: "#/components/schemas/ImportedApiKey"
          title: List of imported keys
          type: array
        next_page_token:
          title: Token for fetching the next page (empty if no more results)
          type: string
      title: ListImportedApiKeysResponse returns paginated list of imported keys
      type: object
    ListIssuedApiKeysResponse:
      example:
        next_page_token: next_page_token
        issued_api_keys:
        - metadata: "{}"
          create_time: 2000-01-23T04:56:07.000+00:00
          rate_limit_policy:
            unit: unit
            quota: quota
            window: window
          visibility: KEY_VISIBILITY_UNSPECIFIED
          key_id: key_id
          expire_time: 2000-01-23T04:56:07.000+00:00
          revocation_description: revocation_description
          update_time: 2000-01-23T04:56:07.000+00:00
          last_used_time: 2000-01-23T04:56:07.000+00:00
          name: name
          revocation_reason: REVOCATION_REASON_UNSPECIFIED
          actor_id: actor_id
          scopes:
          - scopes
          - scopes
          ip_restriction:
            allowed_cidrs:
            - allowed_cidrs
            - allowed_cidrs
          status: KEY_STATUS_UNSPECIFIED
        - metadata: "{}"
          create_time: 2000-01-23T04:56:07.000+00:00
          rate_limit_policy:
            unit: unit
            quota: quota
            window: window
          visibility: KEY_VISIBILITY_UNSPECIFIED
          key_id: key_id
          expire_time: 2000-01-23T04:56:07.000+00:00
          revocation_description: revocation_description
          update_time: 2000-01-23T04:56:07.000+00:00
          last_used_time: 2000-01-23T04:56:07.000+00:00
          name: name
          revocation_reason: REVOCATION_REASON_UNSPECIFIED
          actor_id: actor_id
          scopes:
          - scopes
          - scopes
          ip_restriction:
            allowed_cidrs:
            - allowed_cidrs
            - allowed_cidrs
          status: KEY_STATUS_UNSPECIFIED
      properties:
        issued_api_keys:
          items:
            $ref: "#/components/schemas/IssuedApiKey"
          type: array
        next_page_token:
          type: string
      type: object
...
    SelfRevokeApiKeyRequest:
      description: |-
        SelfRevokeApiKeyRequest allows an API key holder to revoke their own key
        by providing the full key secret as proof of possession.
      example:
        reason: REVOCATION_REASON_UNSPECIFIED
        credential: credential
      properties:
        credential:
          title: Full API key secret or imported key (REQUIRED)
          type: string
        reason:
          $ref: "#/components/schemas/RevocationReason"
      type: object
    SelfRevokeApiKeyResponse:
      description: |-
        SelfRevokeApiKeyResponse is returned on successful self-revocation.

        Empty on success
      type: object
    Status:
      example:
        code: 0
        details:
        - '@type': '@type'
        - '@type': '@type'
        message: message
      properties:
        code:
          format: int32
          type: integer
        details:
          items:
            $ref: "#/components/schemas/Any"
          type: array
        message:
          type: string
      type: object
...